Move nondeterministic extstore logic into internal-non-workflow.
