+++ packages/common/src/internal-non-workflow/index.ts
@@ -6,6 +6,7 @@
 export * from './codec-helpers';
 export * from './codec-types';
 export * from './data-converter-helpers';
+export * from './extstore-helpers';
 export * from './parse-host-uri';
 export * from './proxy-config';
 export * from './tls-config';
